rename component to cardComponent, fix styles, import & use commons component
